Fix big bug with en passant squares
I made the mistake of mixing up an "empty square" (which really
represents the lack of a piece) with a null en passant square. Since
the value of `a1` and `EMPTY_SQUARE` both translate to `0`, this
created an overlap which causes the `a1` square to get overwritten in
certain situations.